Add a parent pointer to egit_object
This keeps us from having to look up the parent in a hash table
whenever an object is finalized. It also lets us deal with more
complicated relationships where e.g. the data stored in a
git_index_entry is owned by the git_index, but the git_index_entry
struct doesn't actually have a pointer to the owner.

This way, we can associate the wrapped git_index to the wrapped
git_index_entry, and ensure that it's released in a timely manner
instead of having to tediously duplicate git_index_entry data
whenever we pass them on to Emacs, so as to ensure that *we* own it.

This will be more useful in the future when more objects are
introduced that share data this way (see: diffs, patches).