Makefile: Add new implementation
This new implementation is based on the same template that I use for
most of my packages.  It is suitable for building the package locally.

The old Emake-based implementation was not suitable for local use,
because it contained no target to fetch the necessary Emake tools,
which is being done in ".travis.yml".  It would have been possible to
fix that, but since I have other reservations about Emake, I decided
otherwise.

Emake currently does not support "make -f MAKEFILE" (actually this is
supposed to be supported but I could not get it to work), so "both
Makefiles" have to be located inside the same actual "Makefile" for
the time being.  The two implementations are separated from each other
using an `ifeq'.

Continue to use Emake on Travis because it supports building on macOS,
which the setup that I usually use hasn't been taught about.  Also the
binaries I currently use for other packages were not build with
support for modules.