Load module silently
Starting with Emacs 27 `load' defaults to showing a message when
loading a module as it is supposed to.  We don't want that so we
have to use `load' directly instead of `load-file' and provide t
for the NOMESSAGE argument, which the latter lacks.  If we did
not do this, then building Magit would be very noisy.