libgit--module-file: hardcode ".so" suffix
As far as I can tell we hardcode the ".so" suffix in our Makefile.
That means we should not rely on the variable `module-file-suffix',
which may have a value other than the expected ".so".

And indeed it has been alleged that the value is different on Darwin
when using Emacs 28 at least.  But because there is nothing to be
gained from using a potentially conflicting value, we simply always
hardcode the suffix, regardless of platform or anything else.

Closes #89.
Closes #98.
Closes #99.

It would probably be better to instead use the appropriate suffix on
all platforms, e.g. ".dylib" on Darwin, but it is unlikely anyone will
put in the work.